Move blueprint models into own file
