+++ app/models/storage/conversions.from-dao.ts
@@ -4,7 +4,7 @@ import {
   ExerciseBlueprint,
   Rest,
   ProgramBlueprint,
-} from '@/models/session-models';
+} from '@/models/blueprint-models';
 import {
   PotentialSet,
   RecordedExercise,
