+++ app/store/current-session/helpers.ts
@@ -1,4 +1,4 @@
-import { ExerciseBlueprintPOJO } from '@/models/session-models';
+import { ExerciseBlueprintPOJO } from '@/models/blueprint-models';
 import { RecordedSetPOJO } from '@/models/session-models';
 import { LocalDateTime } from '@js-joda/core';
 import { match } from 'ts-pattern';
